requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
python-socketio==5.8.0
python-engineio==4.7.1
Werkzeug==2.3.7
python-dateutil==2.8.2
uuid==1.30
bcrypt==4.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
gevent==23.9.1
gevent-websocket==0.10.1
psycopg2-binary==2.9.7